Changements : - Ajout algo Connexité - Optimisation affichage - Optimisation code
Probleme :
- temps de calcul et cout en temps EN H MIN SEC MS à optimiser car ne marche pas bien partout

+++ core/Algo.java
@@ -17,7 +17,7 @@ public abstract class Algo<E extends Comparable<E>> {
     protected int destination;
     protected Label_Generique dest;        // Label destinataire
     protected boolean affichageDeroulementAlgo;      // Affichage du deroulement de l'algo
-    protected long duree;                  // Duree d'execution
+    protected double duree;                  // Duree d'execution
     protected boolean connexes;            // Les noeuds origine et destination sont-ils connexes
 
     protected Algo(Graphe gr) {
